Extract Dungeons from DB.
