+++ src/game_economics/shop_ext.cpp
@@ -526,31 +526,6 @@ void DoStoreShop(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 		SendMsgToChar("Чего хотим? Могу пока только характерстики.\r\n", ch);
 }
 
-
-void RemoveShopSeller(MobRnum mrn) {
-	MobVnum mvn = mob_index[mrn].vnum;
-
-	for (const auto &shop : shop_list) {
-		auto it = std::find(shop->mob_vnums().begin(), shop->mob_vnums().end(), mvn);
-
-		if (it != std::end(shop->mob_vnums())) {
-			shop->remove_mob_vnum(it);
-			mob_index[mrn].func = nullptr;
-		}
-	}
-}
-
-void AddDungeonShopSeller(MobRnum mrn_from, MobRnum mrn_to) {
-	MobVnum mvn_from  = mob_index[mrn_from].vnum;
-
-	for (const auto &shop : shop_list) {
-		if (std::find(shop->mob_vnums().begin(), shop->mob_vnums().end(), mvn_from) != std::end(shop->mob_vnums())) {
-			shop->add_mob_vnum(mob_index[mrn_to].vnum);
-			mob_index[mrn_to].func = shop_ext;
-		}
-	}
-}
-
 void do_shops_list(CharData *ch) {
 /*	DictionaryPtr dic = DictionaryPtr(new Dictionary(SHOP));
 	size_t n = dic->Size();
