+++ src/game_economics/shop_ext.h
@@ -13,8 +13,6 @@
 
 namespace ShopExt {
 
-void RemoveShopSeller(MobRnum mrn);;
-void AddDungeonShopSeller(MobRnum mrn_from, MobRnum mrn_to);
 void do_shops_list(CharData *ch);
 void DoStoreShop(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
 void load(bool reload);
